+++ packages/cardano-services/test/StakePool/TypeormStakePoolProvider/TypeormStakePoolProvider.test.ts
@@ -629,7 +629,7 @@ describe('TypeormStakePoolProvider', () => {
           describe('by default, result is sorted by relevance', () => {
             it('with stringent search', async () => {
               const response = await provider.queryStakePools({ filters: { text: 'sp11' }, pagination });
-              expect(response.pageResults.map(({ metadata }) => metadata?.ticker)).toEqual(['SP11', 'SP10', 'SP1']);
+              expect(response.pageResults.map(({ metadata }) => metadata?.ticker)).toEqual(['SP11', 'SP1', 'SP10']);
             });
 
             it('with mild search', async () => {