+++ packages/cardano-services/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.28.9](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/cardano-services@0.28.8...@cardano-sdk/cardano-services@0.28.9) (2024-06-20)
+
+**Note:** Version bump only for package @cardano-sdk/cardano-services
+
+## [0.28.8](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/cardano-services@0.28.7...@cardano-sdk/cardano-services@0.28.8) (2024-06-18)
+
+**Note:** Version bump only for package @cardano-sdk/cardano-services
+
+## [0.28.7](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/cardano-services@0.28.6...@cardano-sdk/cardano-services@0.28.7) (2024-06-17)
+
+**Note:** Version bump only for package @cardano-sdk/cardano-services
+
+## [0.28.6](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/cardano-services@0.28.5...@cardano-sdk/cardano-services@0.28.6) (2024-06-14)
+
+**Note:** Version bump only for package @cardano-sdk/cardano-services
+
 ## [0.28.5](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/cardano-services@0.28.4...@cardano-sdk/cardano-services@0.28.5) (2024-06-05)
 
 ### Features