+++ packages/dapp-connector/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.12.21](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/dapp-connector@0.12.20...@cardano-sdk/dapp-connector@0.12.21) (2024-06-20)
+
+**Note:** Version bump only for package @cardano-sdk/dapp-connector
+
+## [0.12.20](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/dapp-connector@0.12.19...@cardano-sdk/dapp-connector@0.12.20) (2024-06-14)
+
+**Note:** Version bump only for package @cardano-sdk/dapp-connector
+
 ## [0.12.19](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/dapp-connector@0.12.18...@cardano-sdk/dapp-connector@0.12.19) (2024-06-05)
 
 **Note:** Version bump only for package @cardano-sdk/dapp-connector