+++ packages/governance/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.10.8](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/governance@0.10.7...@cardano-sdk/governance@0.10.8) (2024-06-20)
+
+**Note:** Version bump only for package @cardano-sdk/governance
+
+## [0.10.7](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/governance@0.10.6...@cardano-sdk/governance@0.10.7) (2024-06-14)
+
+**Note:** Version bump only for package @cardano-sdk/governance
+
 ## [0.10.6](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/governance@0.10.5...@cardano-sdk/governance@0.10.6) (2024-06-05)
 
 **Note:** Version bump only for package @cardano-sdk/governance