+++ packages/projection-typeorm/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.22](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/projection-typeorm@0.8.21...@cardano-sdk/projection-typeorm@0.8.22) (2024-06-20)
+
+**Note:** Version bump only for package @cardano-sdk/projection-typeorm
+
+## [0.8.21](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/projection-typeorm@0.8.20...@cardano-sdk/projection-typeorm@0.8.21) (2024-06-18)
+
+**Note:** Version bump only for package @cardano-sdk/projection-typeorm
+
+## [0.8.20](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/projection-typeorm@0.8.19...@cardano-sdk/projection-typeorm@0.8.20) (2024-06-17)
+
+**Note:** Version bump only for package @cardano-sdk/projection-typeorm
+
+## [0.8.19](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/projection-typeorm@0.8.18...@cardano-sdk/projection-typeorm@0.8.19) (2024-06-14)
+
+**Note:** Version bump only for package @cardano-sdk/projection-typeorm
+
 ## [0.8.18](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/projection-typeorm@0.8.17...@cardano-sdk/projection-typeorm@0.8.18) (2024-06-05)
 
 ### Bug Fixes