+++ packages/tx-construction/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.19.3](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/tx-construction@0.19.2...@cardano-sdk/tx-construction@0.19.3) (2024-06-20)
+
+**Note:** Version bump only for package @cardano-sdk/tx-construction
+
+## [0.19.2](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/tx-construction@0.19.1...@cardano-sdk/tx-construction@0.19.2) (2024-06-17)
+
+**Note:** Version bump only for package @cardano-sdk/tx-construction
+
+## [0.19.1](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/tx-construction@0.19.0...@cardano-sdk/tx-construction@0.19.1) (2024-06-14)
+
+**Note:** Version bump only for package @cardano-sdk/tx-construction
+
 ## [0.19.0](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/tx-construction@0.18.7...@cardano-sdk/tx-construction@0.19.0) (2024-06-05)
 
 ### ⚠ BREAKING CHANGES